reorganize functions logically
